FIX: Display browse more message as long as there are suggested topics
