+++ app/assets/javascripts/discourse/app/components/suggested-topics.hbs
@@ -23,8 +23,4 @@
       <BasicTopicList @topics={{this.topic.suggestedTopics}} />
     {{/if}}
   </div>
-
-  <h3 class="suggested-topics-message">
-    {{html-safe this.browseMoreMessage}}
-  </h3>
 </div>
