+++ app/assets/stylesheets/common/base/topic.scss
@@ -369,27 +369,31 @@ a.badge-category {
 }
 
 // Target the .badge-category text, the bullet icon needs to maintain `display: block`
-.more-content-topics h3 .badge-wrapper.bullet span.badge-category,
-.more-content-topics h3 .badge-wrapper.box span,
-.more-content-topics h3 .badge-wrapper.bar span {
+.suggested-topics-message .badge-wrapper.bullet span.badge-category,
+.suggested-topics-message .badge-wrapper.box span,
+.suggested-topics-message .badge-wrapper.bar span {
   display: inline;
 }
 
-.more-content-topicss h3 .badge-wrapper.bullet span.badge-category {
+.suggested-topics-message .badge-wrapper.bullet span.badge-category {
   // Override vertical-align: text-top from `badges.css.scss`
   vertical-align: baseline;
   line-height: var(--line-height-medium);
 }
 
-.more-content-topics h3 .badge-wrapper.bullet,
-.more-content-topics h3 .badge-wrapper.bullet span.badge-category-parent-bg,
-.more-content-topics h3 .badge-wrapper.bullet span.badge-category-bg {
+.suggested-topics-message .badge-wrapper.bullet {
+  margin-right: 0;
+}
+
+.suggested-topics-message .badge-wrapper.bullet,
+.suggested-topics-message .badge-wrapper.bullet span.badge-category-parent-bg,
+.suggested-topics-message .badge-wrapper.bullet span.badge-category-bg {
   // Top of bullet aligns with top of line - adjust line height to vertically align bullet.
   line-height: 0.8;
 }
 
-.more-content-topics .badge-wrapper.bullet span.badge-category,
-.more-content-topics .badge-wrapper.bar span.badge-category {
+.suggested-topics-message .badge-wrapper.bullet span.badge-category,
+.suggested-topics-message .badge-wrapper.bar span.badge-category {
   max-width: 150px;
 }
