+++ app/assets/stylesheets/desktop/topic.scss
@@ -146,8 +146,7 @@
 
     .topic-list-header,
     .posts-map,
-    .views,
-    .suggested-topics-message {
+    .views {
       display: none;
     }
   }
